+++ src/classes/libraries/ColorLibrary.java
