+++ src/classes/tools/ColoringTools.java
@@ -8,7 +8,6 @@ import java.util.List;
 
 // Import Custom Libraries
 import classes.libraries.MethodLibrary;
-import classes.libraries.ConstLibrary;
 
 public class ColoringTools implements MethodLibrary
 {
